requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.10.0
gunicorn==19.7.1
httpie==1.0.2
idna==2.8
numpy==1.16.4
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
scipy==1.3.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==3.3.1
